requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
supabase==2.13.0
pydantic==2.10.6
jinja2==3.1.5
